+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - Untemplated 'Command' job for pyrotorque
 ## Fixed
 - Show better message when rtxmlrpc gets no method name
+- Allow `mktor` and `chtor` to run without rtorrent.rc
 ## Removed
 - Theming module